+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## Gazebo 7.x.x (2018-xx-xx)
 
+1. Process insertions and deletions on gz log echo
+    * [Pull request 2608](https://bitbucket.org/osrf/gazebo/pull-request/2608)
+    * [Issue 2136](https://bitbucket.org/osrf/gazebo/issues/2136)
+
 
 ## Gazebo 7.10.0 (2018-02-07)